optimal plan for road,arranged basics in order
